git-svn-id: https://svn.code.sf.net/p/sidplay-residfp/code/branches/libsidplayfp-1.6@1735 75748be7-0387-4326-94c7-f41ce0883d2c

+++ sidplayfp/c64/VIC_II/sprites.h
@@ -156,7 +156,7 @@ public:
      *
      * @param val sprites bitmask
      */
-    bool isDma(uint8_t val)
+    bool isDma(unsigned int val)
     {
         return dma && val;
     }
